requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
scipy
umap-learn
